Twitch streamer live tracker (SI) (#3)
* Changed IPlatformStreamerLiveTracker methods to simply work with strings (usernames of streamers).
Changed IStreamersLiveStatusProvider to take callbacks which take StreamStatus as param instead of just Streamer. Provides more info to callbacks with StreamStatus as argument.

* (wip) First implementation of twitch streamer live tracker using setInterval for polling. Along with older version of tracker using setTimeout that I abandoned due to concurrency issues.

* (wip) Formatting of twitch live tracker along with implmenting updateResolvedStreamerCache and fixing addStreamer to check if the streamer already exists. Also, ensures pipeline of add streamer to internal tracking set, emit add event, then emit livestatus checks if fast-check was performed.

* (needs testing) Refactor of twitch live tracker implementation. Should be complete for the most part but needs to be tested.

* Simplified add and remove streamer implementations as the previous ones created a lot of strange corner cases for users of the interface with things like, if within an add resolving the streamer name failed it still kept the name within the trackedStreamer set. Also, the "optimizations" of being able to hold off till the tracker is started to then make a batched api call was ruined by the fact that emitting an add event required an immediate api call if the streamer wasn't already resolved. The whole thing with "fast-checks" also just added more complexity that is unneeded for now.
Add implementation simply uses the retrieveValidResolvedStreamers first as a way to check if the streamer names already exists as a valid entry in the resolved streamer cache, and if not immediately makes an api call to add them to the resolved cache, but also provides a guard of ensuring the streamer being added is a valid twtich user. Then they are added and the "safeEmit" is reverted back to a simple synchronous emitModifyingEvent as it is only invoked after a Streamer object has been obtained.
removeStreamer now just pulls the streamer object directly from the resolvedStreamer cache and emits the event that it was removed, as it now enjoys from the trackedStreamer set being guarded to only valid streamers. In the previous implementation, if an invalid streamer was added to the set, an error would always be thrown when trying to remove them from the set as a safeEmit necessarily required a valid streamer name.

* Included an info log message for adding a streamer to begin tracking.